Update automatically when a new hike is entered
To do this is a little convoluted because right now I have the state for
the hike list in the parent and the state for the submission fields in
that sub component.  In the future I would like to refactor so that all
the state lives in the parent container.  Right now I have to pass down
a bound version of my addHike which makes the ajax post request and the
function to set the state.  To get it to work correctly I need to call
the addHike function in the submission component with the the submission
component's state as the data for the post request and the bound version
of the setState function.  The addHike function makes the post request
and then calls the setState function.  If I had the state of everything
in my parent I would only need to pass functions down to the submission
component that deals with when the state changes, but is that any
better?