Solves the connection issues locally
I cant see the process.env variables locally because that is a heroku.
So I added back in my local dabatase in the url for connecting to the
database as the second argument after an or logical operator.  Also I
was using the wrong process.envir so I am hoping that solves my
deployment issues. We will see!