Updates existing entry if duplicate detected
Now if you input a hike with a name that already exists it will update
the time so that the most recent one is showing.  This satisfies my
update criteria for a CRUD app.  Next I want to add a link to each entry
that will only delete that entry.  If I have time I want to store the
data differently so that the user can decide how to view the data.  The
options will be best, most recent, or all.