Adds an option to delete a single entry
I do this by creating the function in the parent container and pass a
bounded version of it down to its grandchild.  There I use it when a
delete entry link is clicked.  Each link has a class name of their hike
and I can see that class name in the grandparent using the
e.target.className.  Then I use that class name in a post request to my
server handle the request and call the database function that takes the
hike name and deletes that entry and sends back the new list of hikes
which the server in turn sends to the view which sets the state for the
new info.