Adds a location field
It adds a new input field on the view with all the necessary state and
state changing functions.  It gets sent to my server as data and I split
out the three fields into new varibales.  Send those variables to my
database and use the arguemnts to add something or update an entry in my
database.